[2/2]使用clippy优化代码

+++ capdows/src/i18n.rs
@@ -26,6 +26,6 @@ impl LangID {
         }
     }
     pub const fn split(self) -> (u8, u16) {
-        ((self.id.get() >> 10) as u8, (self.id.get() & 0x3FF) as u16)
+        ((self.id.get() >> 10) as u8, (self.id.get() & 0x3FF))
     }
 }
